+++ entities/image/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @brains/image-plugin
 
+## 0.2.0-alpha.411
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.411
+  - @brains/image@0.2.0-alpha.411
+  - @brains/utils@0.2.0-alpha.411
+  - @brains/entity-service@0.2.0-alpha.411
+  - @brains/plugins@0.2.0-alpha.411
+
+## 0.2.0-alpha.410
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.410
+  - @brains/image@0.2.0-alpha.410
+  - @brains/utils@0.2.0-alpha.410
+  - @brains/entity-service@0.2.0-alpha.410
+  - @brains/plugins@0.2.0-alpha.410
+
 ## 0.2.0-alpha.409
 
 ### Patch Changes