+++ interfaces/chat/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @brains/chat
 
+## 0.2.0-alpha.411
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.411
+  - @brains/utils@0.2.0-alpha.411
+  - @brains/auth-service@0.2.0-alpha.411
+  - @brains/plugins@0.2.0-alpha.411
+
+## 0.2.0-alpha.410
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.410
+  - @brains/utils@0.2.0-alpha.410
+  - @brains/auth-service@0.2.0-alpha.410
+  - @brains/plugins@0.2.0-alpha.410
+
 ## 0.2.0-alpha.409
 
 ### Patch Changes