+++ plugins/admin/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @brains/admin
 
+## 0.2.0-alpha.411
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/utils@0.2.0-alpha.411
+  - @brains/auth-service@0.2.0-alpha.411
+  - @brains/plugins@0.2.0-alpha.411
+
+## 0.2.0-alpha.410
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/utils@0.2.0-alpha.410
+  - @brains/auth-service@0.2.0-alpha.410
+  - @brains/plugins@0.2.0-alpha.410
+
 ## 0.2.0-alpha.409
 
 ### Patch Changes