+++ shared/eslint-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @brains/eslint-config
 
+## 0.2.0-alpha.411
+
+## 0.2.0-alpha.410
+
 ## 0.2.0-alpha.409
 
 ## 0.2.0-alpha.408