+++ shared/media-page-composer/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @brains/media-page-composer
 
+## 0.2.0-alpha.411
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.411
+  - @brains/media-renderer@0.2.0-alpha.411
+  - @brains/site-engine@0.2.0-alpha.411
+  - @brains/ui-library@0.2.0-alpha.411
+  - @brains/utils@0.2.0-alpha.411
+  - @brains/plugins@0.2.0-alpha.411
+
+## 0.2.0-alpha.410
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.410
+  - @brains/media-renderer@0.2.0-alpha.410
+  - @brains/site-engine@0.2.0-alpha.410
+  - @brains/ui-library@0.2.0-alpha.410
+  - @brains/utils@0.2.0-alpha.410
+  - @brains/plugins@0.2.0-alpha.410
+
 ## 0.2.0-alpha.409
 
 ### Patch Changes