+++ shared/operation-context/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @brains/operation-context
 
+## 0.2.0-alpha.411
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.411
+
+## 0.2.0-alpha.410
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.410
+
 ## 0.2.0-alpha.409
 
 ### Patch Changes