+++ sites/default/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @brains/site-default
 
+## 0.2.0-alpha.411
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/site-professional@0.2.0-alpha.411
+
+## 0.2.0-alpha.410
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/site-professional@0.2.0-alpha.410
+
 ## 0.2.0-alpha.409
 
 ### Patch Changes